refactor: rename adapter package structure for consistency across controllers, mappers, presenters, and repositories
