+++ app/api/src/main/java/com/somba/api/adapter/controllers/GlobalExceptionHandler.java
@@ -1,12 +1,13 @@
-package com.somba.api.adapters.controllers;
+package com.somba.api.adapter.controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import com.somba.api.adapter.presenters.ErrorDetails;
+
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
-
-import com.somba.api.adapters.presenters.ErrorDetails;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
